requirements.txt
fpdf2
pathlib
PIL
streamlit
pdf2docx
openai
pdfminer
docx
tiktoken